Use DcMotorEx for all motors and update the SuperStructure class

+++ FtcRobotController-master/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/utils/Positions.java
@@ -3,10 +3,10 @@ package org.firstinspires.ftc.teamcode.utils;
 import androidx.annotation.NonNull;
 
 public class Positions {
-    public int left;
-    public int right;
+    public double left;
+    public double right;
 
-    public Positions(int left, int right) {
+    public Positions(double left, double right) {
         this.left = left;
         this.right = right;
     }
